requirements.txt
asgiref==3.2.3
Django==3.0.2
django-bootstrap3==12.0.3
lxml==4.5.0
python-docx==0.8.10
pytz==2019.3
sqlparse==0.3.0
